Optimizing sha1sum check in the downloading script (#6)
Summary:
The current downloading script performs a sha1sum check after the whole file has been downloaded. For that, it re-reads the file from disk again. However, the sha1sum can be computed during the download, significantly speeding up the process. This pull request introduces precisely that behavior. It was tested by successfully downloading several sequences of the dataset. As a minor fix, it also fixes a typo in logger.info ("Donwload" -> "Download").

Pull Request resolved: https://github.com/facebookresearch/nymeria_dataset/pull/6

Reviewed By: yutingye

Differential Revision: D64766495

fbshipit-source-id: 289c8869f42d1cff5cbce3007547b74bb73812c5